refactor(engine): Extract Web Speech logic into a default SpeechEngine
Split the monolithic Vocal factory into an engine-agnostic core and a
default `SpeechEngine` (the Web Speech implementation):

- `src/types.ts` holds the shared types, the `eventTypes` map and the new
  `SpeechEngineInstance` / `SpeechEngineContext` / `SpeechEngineFactory`
  contract (not yet publicly exported).
- `src/SpeechEngine.ts` owns everything backend-specific: vendor
  resolution, recognition setup, the continuous restart loop, fatal-error
  handling, final-result aggregation and the permission watch. It pushes
  already-shaped events to the core through `ctx.emit`.
- `src/Vocal.ts` becomes a thin core: user listener registry, event
  fan-out, the `isRecording` getter and lifecycle delegation.

Public API and observable behaviour are unchanged; the engine is hardcoded
to the default for now. Two white-box tests that asserted listener
placement on the native instance are rewritten as behavioural checks, since
user subscription now lives in the core.
